refactor: separate the components
