+++ src/views/admin/home/styled.js
@@ -1,0 +1,5 @@
+import styled from 'styled-components'
+
+export const ContainerChart = styled.div`
+  margin-top: 5%;
+`
